requirements.txt
streamlit==1.32.0
watchdog==4.0.0
folium==0.16.0
streamlit-folium==0.18.0
pandas==2.2.0
cryptography==42.0.0
Pillow==10.2.0
